+++ site/gatsby-site/playwright/e2e/pages.1.spec.ts
@@ -1,0 +1,15 @@
+import { test } from '../utils';
+import { testPages } from './pages-tests';
+
+test.describe('Pages 1', () => {
+
+  const paths = [
+    '/',
+    '/taxonomies/',
+    '/apps/discover/',
+    '/apps/incidents/',
+    '/apps/newsdigest/',
+  ];
+
+  testPages(paths);
+});
