+++ site/gatsby-site/playwright/e2e/pages.2.spec.ts
@@ -1,0 +1,15 @@
+import { test } from '../utils';
+import { testPages } from './pages-tests';
+
+test.describe('Pages 2', () => {
+
+  const paths = [
+    '/apps/submit/',
+    '/apps/submitted/',
+    '/apps/variants/',
+    '/blog/',
+    '/summaries/flagged/',
+  ];
+
+  testPages(paths);
+});
